requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.38.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-mailjet==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
requests==2.31.0
